+++ Source/Krypton Toolkit/Main/Krypton.Toolkit.Suite.Extended.Notifications/UX/Toast/Version 2/KryptonToastNotificationVersion2.cs
@@ -1,0 +1,20 @@
+﻿#region BSD License
+/*
+ * Use of this source code is governed by a BSD-style
+ * license or other governing licenses that can be found in the LICENSE.md file or at
+ * https://raw.githubusercontent.com/Krypton-Suite/Extended-Toolkit/master/LICENSE
+ */
+#endregion
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Krypton.Toolkit.Suite.Extended.Notifications
+{
+    class KryptonToastNotificationVersion2
+    {
+    }
+}
